client: rename `connect(config, server)` of mobile
To have a unify flow to initialize VPN network, the funcion renamed as
the same cli using.
`mobile::lightway::lightway_client_connect(..)` is moved to
`connect(ClientConfig, ClientConnectionConfig)` with the same
function name and function inputs.
