+++ lightway-client/src/config.rs
@@ -19,6 +19,14 @@ use struct_patch::Patch;
 use std::net::IpAddr;
 
 // NOTE
+// These internal parameters are not exposed
+#[cfg(mobile)]
+mod internal {
+    pub(crate) const INTERNAL_MTU: u16 = 1350;
+    pub(crate) const MAX_SOCKET_BUFFER_LEN: usize = 1024000;
+}
+
+// NOTE
 // The cli argument and options all set to ConfigPatch, which is generated by
 // Patch derive, do NOT set defualts in the clap macros, and let all default
 // values follow the Rust convention in Default trait, such that we are able
@@ -439,12 +447,16 @@ impl Config {
             tun_local_ip,
             tun_dns_ip,
             preferred_connection_wait_interval,
+            enable_pmtud,
             ..
         } = self;
         crate::ClientConfig {
             tun_local_ip,
             tun_dns_ip,
             preferred_connection_wait_interval: preferred_connection_wait_interval.into(),
+            internal_mtu: internal::INTERNAL_MTU,
+            max_socket_buffer_len: internal::MAX_SOCKET_BUFFER_LEN,
+            enable_pmtud,
 
             tun_fd,
             external_event_handler,
